fix bug.

+++ CrystalSpace/CrystalSpaceView/MarchingCubesView.cpp
@@ -32,9 +32,9 @@ void MarchingCubesView::onOk()
 
 	const auto center = box.getCenter();
 	const auto radius = 5.0;
-	for (size_t i = 0; i < 100; ++i) {
-		for (size_t j = 0; j < 100; ++j) {
-			for (size_t k = 0; k < 100; ++k) {
+	for (int i = 0; i < 100; ++i) {
+		for (int j = 0; j < 100; ++j) {
+			for (int k = 0; k < 100; ++k) {
 				const auto distanceSquared = Math::getDistanceSquared(v.getPositionAt({ i, j, k }), center);
 				if (distanceSquared < radius * radius) {
 					const auto distance = ::sqrt(distanceSquared);
@@ -47,6 +47,7 @@ void MarchingCubesView::onOk()
 		}
 	}
 
+	/*
 	MarchingCubesAlgo algo;
 	algo.build(v, 0.30f);
 	const auto& triangles = algo.getTriangles();
@@ -64,4 +65,5 @@ void MarchingCubesView::onOk()
 	//builder.pushCurrentFaceGroup();
 	//getWorld()->getObjectFactory()->createPolygonMeshScene(builder.(), "MarchingCubes");
 	//getWorld()->updateViewModel();
+	*/
 }
